+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Updated Readme.md
 - Added onFinished to the index.d.ts file (TS support).
+- Request cancel update.
 
 ## v0.0.9
 - Added option to override the config of an request.